yinsu/3dmaze/requirements.txt
# requirements.txt
# This file is for users who cannot use Conda.
# Note: For GPU support, PyTorch must be installed separately by following the
# official instructions at pytorch.org.

# For PyTorch (GPU), run this command first:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

gymnasium
stable-baselines3
memory-maze
pyyaml
tensorboard
opencv-python
numpy